数禾

See merge request pd_str/marketing!3

+++ marketing-cs/src/main/java/com/br/marketing/entity/TaskStatus.java
@@ -16,6 +16,6 @@ public class TaskStatus {
     private String incrDate;
     private Integer incrStatus;
     private Integer onceStatus;
-    private Integer fileId;
+    private Long fileId;
 
 }
